Validation of the CSRF token has been improved - admin (first part)

+++ modules/Settings/Companies/actions/DeleteAjax.php
@@ -16,22 +16,13 @@ class Settings_Companies_DeleteAjax_Action extends Settings_Vtiger_Delete_Action
 	 */
 	public function process(\App\Request $request)
 	{
-		$record = $request->get('record');
-		$qualifiedModuleName = $request->getModule(false);
-		$recordModel = Settings_Companies_Record_Model::getInstance($record);
+		$recordModel = Settings_Companies_Record_Model::getInstance($request->getInteger('record'));
+		if ($recordModel->get('default') === 1) {
+			throw new \App\Exceptions\NoPermittedToRecord('LBL_PERMISSION_DENIED');
+		}
 		$recordModel->delete();
-
-		$moduleModel = Settings_Vtiger_Module_Model::getInstance($qualifiedModuleName);
-		header("Location: {$moduleModel->getDefaultUrl()}");
-	}
-
-	/**
-	 * Validate Request.
-	 *
-	 * @param \App\Request $request
-	 */
-	public function validateRequest(\App\Request $request)
-	{
-		$request->validateReadAccess();
+		$response = new Vtiger_Response();
+		$response->setResult(Settings_Vtiger_Module_Model::getInstance($request->getModule(false))->getDefaultUrl());
+		$response->emit();
 	}
 }
